image_pipeline/requirements.txt
accelerate==0.31.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-eventhub==5.12.1
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-mgmt-compute==31.0.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.4.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.20.0
azure-storage-queue==12.10.0
backcall==0.2.0
bitarray==2.9.2
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1717618050233/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
crcmod==1.7
cryptography @ file:///croot/cryptography_1714660666131/work
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.5.14
datasets==2.18.0
debugpy==1.8.2
decorator==5.1.1
decord==0.6.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fairseq==0.12.2
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gast==0.6.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
hydra-core==1.0.7
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
isodate==0.6.1
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-text-splitters==0.2.2
langsmith==0.1.83
lazy_loader==0.4
librosa==0.10.2.post1
llama-index @ git+https://github.com/hex-plex/llama_index@d7a34f38673d39a2af25264d0445505260e8b525
llvmlite==0.41.1
lxml==5.2.2
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.21.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
modelscope==1.15.0
mpmath @ file:///croot/mpmath_1690848262763/work
msal==1.29.0
msal-extensions==1.2.0
msgpack==1.0.8
msrest==0.7.1
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1690561992265/work
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.0.6
open-clip-torch==2.24.0
openai==1.35.10
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.6
oss2==2.18.6
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.10.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==26.0.3
rapidfuzz==3.9.4
regex==2024.5.15
requests @ file:///croot/requests_1716902831423/work
requests-oauthlib==2.0.0
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.31
stack-data==0.6.3
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
triton==2.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.48
ultralytics-thop==2.0.0
unicodedata2==15.1.0
urllib3==1.26.19
wcwidth==0.2.13
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zhconv==1.4.3
zipp==3.19.2
